requirements.txt
python-dotenv>=1.0.0
pexpect>=4.9
